# [perf] Mark hottest functions as `#[inline]` to help the optimizer