Before:

```
  Time (mean ± σ):     783.7 ms ±  15.9 ms    [User: 763.8 ms, System: 20.0 ms]
  Range (min … max):   763.6 ms … 808.8 ms    10 runs
```

After:

```
  Time (mean ± σ):     770.8 ms ±   4.7 ms    [User: 749.8 ms, System: 21.0 ms]
  Range (min … max):   762.6 ms … 777.6 ms    10 runs
```

That's 98.35% of the previous time.

```
  Instructions:          5850997131 (-2.919522%)
  L1 Accesses:           7725477299 (-3.742230%)
  Estimated Cycles:      8082559469 (-3.582845%)
```